# Enable split rotary fusion for batched model